# Conflicts:
#	cockroach/main.py

+++ cassandra/transactions/t3.py
@@ -1,11 +1,8 @@
-from cassandra.query import named_tuple_factory, SimpleStatement
+from cassandra.query import BatchStatement, SimpleStatement
 from decimal import Decimal
 
-# todo: remove all prints and queries for checking if updates are correctly applied before testing.
 def execute_t3(session, args_arr):
-    # D,2,8
-    print("T3 Delivery Transaction called!\n----------------------")
-    # print(args_arr)
+    print("T3 Delivery Transaction called!")
 
     ### extract query inputs    
     assert len(args_arr) == 3, "Wrong length of argments for T3"
@@ -20,31 +17,32 @@ def execute_t3(session, args_arr):
       - C_DELIVERY_CNT, in tables: customer
     """
     
-    # for d_id in range(1, 2):
     for d_id in range(1, 11): # [1..10]
       # Orders table: (o_w_id, o_d_id, o_id) uniquely identify row
       # So given (o_w_id, o_d_id), each o_id returned should be unique
       # Expect: bc of clustering, smallest o_id is first.
-      order = session.execute(f"""
+      orders = session.execute(f"""
         SELECT * FROM order_by_carrier_id WHERE o_w_id=%s and o_d_id=%s and o_carrier_id=%s LIMIT 1;""",
-        (w_id, d_id, -1))[0]
-      # assert order, f"No null carrier_id exists for w={w_id}, d={d_id}. Is this a problem or can silently skip?"  # ask
-      if not order:
-        print(f'No null carrier_id exists for w={w_id}, d={d_id}, skipping to next district.')
+        (w_id, d_id, -1))
+      if not orders:
+        # print(f'No null carrier_id exists for w={w_id}, d={d_id}, skipping to next district.')
         continue
-      print(order)
+      order = orders[0]
+      # print(order)
 
       ##### update carrier_id in order_by_carrier_id, which is a PK col
-      session.execute(SimpleStatement(f"""
+      batch = BatchStatement()
+      batch.add(SimpleStatement(f"""
         DELETE FROM order_by_carrier_id 
         WHERE o_w_id={order.o_w_id} 
           AND o_d_id={order.o_d_id} 
           AND o_carrier_id={order.o_carrier_id} 
           AND o_id={order.o_id} 
           AND o_c_id={order.o_c_id};"""))
-      
-      session.execute(f"""
-        INSERT INTO order_by_carrier_id (o_w_id,
+
+      batch.add(SimpleStatement(f"""
+        INSERT INTO order_by_carrier_id (
+                            o_w_id,
                             o_d_id,
                             o_carrier_id,
                             o_id,
@@ -55,8 +53,8 @@ def execute_t3(session, args_arr):
                             o_c_middle,
                             o_entry_d,
                             o_ol_cnt) 
-                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);""", 
-                          (order.o_w_id,
+                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);"""), 
+                         (order.o_w_id,
                           order.o_d_id,
                           new_carrier_id,
                           order.o_id,
@@ -67,26 +65,30 @@ def execute_t3(session, args_arr):
                           order.o_c_middle,
                           order.o_entry_d,
                           order.o_ol_cnt))
+      
+      session.execute(batch)
 
       #### check update of carrier_id is correct
-      chk = session.execute(f"""
-        SELECT * FROM order_by_carrier_id 
-        WHERE o_w_id=%s 
-        and o_d_id=%s 
-        and o_carrier_id=%s 
-        and o_id=%s LIMIT 1;""",
-        (order.o_w_id, order.o_d_id, new_carrier_id, order.o_id))[0]
-      print(chk)
+      # skipping check for benchmarking
+      # chk = session.execute(f"""
+      #   SELECT * FROM order_by_carrier_id 
+      #   WHERE o_w_id=%s 
+      #   and o_d_id=%s 
+      #   and o_carrier_id=%s 
+      #   and o_id=%s LIMIT 1;""",
+      #   (order.o_w_id, order.o_d_id, new_carrier_id, order.o_id))[0]
+      # print(chk)
 
       ##### update carrier_id in orders, which is a PK col
-      session.execute(SimpleStatement(f"""
+      batch = BatchStatement()
+      batch.add(SimpleStatement(f"""
         DELETE FROM orders 
         WHERE o_w_id={order.o_w_id} 
           AND o_d_id={order.o_d_id} 
           AND o_id={order.o_id} 
           AND o_c_id={order.o_c_id};"""))
-      
-      session.execute(f"""
+
+      batch.add(SimpleStatement(f"""
         INSERT INTO orders (o_w_id,
                             o_d_id,
                             o_carrier_id,
@@ -98,8 +100,8 @@ def execute_t3(session, args_arr):
                             o_c_middle,
                             o_entry_d,
                             o_ol_cnt) 
-                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);""", 
-                          (order.o_w_id,
+                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);"""), 
+                         (order.o_w_id,
                           order.o_d_id,
                           new_carrier_id,
                           order.o_id,
@@ -110,18 +112,20 @@ def execute_t3(session, args_arr):
                           order.o_c_middle,
                           order.o_entry_d,
                           order.o_ol_cnt))
-
+      session.execute(batch)      
+      
       #### update order_line table
       # when queried with (ol_w_id, ol_d_id, ol_o_id), each rows return should have unique ol_number
-      print()
       order_lines = session.execute(f"""
           SELECT * FROM order_line WHERE ol_w_id=%s AND ol_d_id=%s AND ol_o_id=%s;""",
           (w_id, d_id, order.o_id))
+      
+      if not order_lines:
+        continue
 
       # note: order_lines is an iterator and will be exhausted. Only one iteration is allowed without duplicating it.
       ol_amount_total = Decimal(0)
       for r in order_lines:
-        print(r)
         ol_amount_total += r.ol_amount
 
         session.execute(f"""
@@ -136,28 +140,34 @@ def execute_t3(session, args_arr):
           (r.ol_w_id, r.ol_d_id, r.ol_o_id, r.ol_quantity, r.ol_number, r.ol_c_id))
         
       #### check order_lines correctly updated
-      print()
-      order_lines = session.execute(f"""
-        SELECT * FROM order_line WHERE ol_w_id=%s AND ol_d_id=%s AND ol_o_id=%s;""",
-        (w_id, d_id, order.o_id))
-      for r in order_lines:
-        print(r)
+      # print()
+      # order_lines = session.execute(f"""
+      #   SELECT * FROM order_line WHERE ol_w_id=%s AND ol_d_id=%s AND ol_o_id=%s;""",
+      #   (w_id, d_id, order.o_id))
+      # for r in order_lines:
+      #   print(r)
 
       #### update customer's C_BALANCE, C_DELIVERY_CNT via drop and insert
-      print('\ntotal', ol_amount_total)
+      # not expected to fail. `order` exists at this point.
       customer = session.execute(SimpleStatement(f"""
         SELECT * FROM customer 
         WHERE c_w_id={w_id} 
         and c_d_id={d_id} 
-        and c_id={order.o_c_id};"""))[0]
-      print(customer)
+        and c_id={order.o_c_id};"""))
+      if not customer:
+        continue
+      customer = customer[0]
+      # print(customer)
 
-      session.execute(SimpleStatement(f"""
+
+      batch = BatchStatement()
+      batch.add(SimpleStatement(f"""
         DELETE FROM customer 
         WHERE c_w_id={w_id} 
         and c_d_id={d_id} 
         and c_id={order.o_c_id};"""))
-      session.execute(f"""
+
+      batch.add(SimpleStatement(f"""
         INSERT INTO customer (c_w_id, 
                               c_d_id, 
                               c_id, 
@@ -181,8 +191,8 @@ def execute_t3(session, args_arr):
                               c_w_name,
                               c_ytd_payment,
                               c_zip) 
-                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);""", 
-                          (customer.c_w_id, 
+                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);"""), 
+                           (customer.c_w_id, 
                             customer.c_d_id, 
                             customer.c_id, 
                             customer.c_balance + ol_amount_total, 
@@ -205,16 +215,19 @@ def execute_t3(session, args_arr):
                             customer.c_w_name,
                             customer.c_ytd_payment,
                             customer.c_zip))
+      
+      session.execute(batch)
 
       #### customer's C_BALANCE duplicated in top_balance table
-      session.execute(SimpleStatement(f"""
+      batch = BatchStatement()
+      batch.add(SimpleStatement(f"""
         DELETE FROM top_balance 
         WHERE c_w_id={customer.c_w_id} 
         and c_d_id={customer.c_d_id} 
         and c_balance={customer.c_balance} 
         and c_id={customer.c_id};"""))
-        
-      session.execute(f"""
+
+      batch.add(SimpleStatement(f"""
         INSERT INTO top_balance (c_w_id,
                                  c_d_id,
                                  c_balance,
@@ -224,7 +237,7 @@ def execute_t3(session, args_arr):
                                  c_last,
                                  c_middle,
                                  c_w_name) 
-                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s);""", 
+                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s);"""), 
                           (customer.c_w_id, 
                            customer.c_d_id, 
                            customer.c_balance + ol_amount_total,
@@ -234,22 +247,21 @@ def execute_t3(session, args_arr):
                            customer.c_last,
                            customer.c_middle,
                            customer.c_w_name))
-
+      session.execute(batch)
 
       #### check customer updated
-      customer = session.execute(SimpleStatement(f"""SELECT * FROM customer WHERE c_w_id={w_id} and c_d_id={d_id} and c_id={order.o_c_id};"""))[0]
-      print(customer)
+      # customer = session.execute(SimpleStatement(f"""SELECT * FROM customer WHERE c_w_id={w_id} and c_d_id={d_id} and c_id={order.o_c_id};"""))[0]
+      # print(customer)
 
 
       #### O_CARRIER_ID duplicated in order_status table
-      print()
-      order_status_before = session.execute(f"""
-        SELECT * FROM order_status where O_W_ID=%s and O_D_ID=%s and O_C_ID=%s and O_ID=%s;""",
-        (w_id, d_id, order.o_c_id, order.o_id))[0]
-      print(order_status_before)
-      assert order_status_before.o_carrier_id == -1, f"Inconsistency detected? Trying to update null order_status carrier_id but it's already non-null" # ask
+      # order_status_before = session.execute(f"""
+      #   SELECT * FROM order_status where O_W_ID=%s and O_D_ID=%s and O_C_ID=%s and O_ID=%s;""",
+      #   (w_id, d_id, order.o_c_id, order.o_id))[0]
+      # print(order_status_before)
+      # assert order_status_before.o_carrier_id == -1, f"Inconsistency detected? Trying to update null order_status carrier_id but it's already non-null" # ask
 
-      # todo: only need to retain this UPDATE query. The ones before and after are for checking while developing
+      # Actual UPDATE query. The ones before and after are for checking while developing
       session.execute(f"""
         UPDATE order_status 
         SET o_carrier_id=%s 
@@ -259,7 +271,7 @@ def execute_t3(session, args_arr):
         and O_ID=%s;""",
         (new_carrier_id, w_id, d_id, order.o_c_id, order.o_id))
 
-      order_status_after = session.execute(f"""
-        SELECT * FROM order_status where O_W_ID=%s and O_D_ID=%s and O_C_ID=%s and O_ID=%s;""",
-        (w_id, d_id, order.o_c_id, order.o_id))[0]
-      print(order_status_after)
+      # order_status_after = session.execute(f"""
+      #   SELECT * FROM order_status where O_W_ID=%s and O_D_ID=%s and O_C_ID=%s and O_ID=%s;""",
+      #   (w_id, d_id, order.o_c_id, order.o_id))[0]
+      # print(order_status_after)
